Converts Statement to tsx
